requirements.txt
requests
rich
fpdf2
python-nmap
fastapi
uvicorn
gevent
gunicorn
pydantic
python-multipart
pyjwt
passlib
bcrypt
email-validator
argon2-cffi
python-dotenv
psycopg2-binary
slowapi
alembic
sqlalchemy
